refactor(rust): More `polars-io` cleanup (#15885)
