+++ crates/polars-io/src/avro/write.rs
@@ -1,10 +1,13 @@
+use std::io::Write;
+
 pub use arrow::io::avro::avro_schema::file::Compression;
 use arrow::io::avro::avro_schema::{self};
 use arrow::io::avro::write;
 use polars_core::error::to_compute_err;
+use polars_core::prelude::*;
 pub use Compression as AvroCompression;
 
-use super::*;
+use crate::shared::SerWriter;
 
 /// Write a [`DataFrame`] to [Apache Avro] format
 ///
