+++ crates/polars-io/src/cloud/mod.rs
@@ -3,22 +3,19 @@
 #[cfg(feature = "cloud")]
 mod adaptors;
 #[cfg(feature = "cloud")]
-pub use adaptors::*;
-
-#[cfg(feature = "cloud")]
-mod polars_object_store;
-#[cfg(feature = "cloud")]
-pub use polars_object_store::*;
-
-#[cfg(feature = "cloud")]
 mod glob;
 #[cfg(feature = "cloud")]
-pub use glob::*;
+mod object_store_setup;
+pub mod options;
+#[cfg(feature = "cloud")]
+mod polars_object_store;
 
 #[cfg(feature = "cloud")]
-mod object_store_setup;
+pub use adaptors::*;
+#[cfg(feature = "cloud")]
+pub use glob::*;
 #[cfg(feature = "cloud")]
 pub use object_store_setup::*;
-
-pub mod options;
 pub use options::*;
+#[cfg(feature = "cloud")]
+pub use polars_object_store::*;
