+++ crates/polars-io/src/csv/read/parser.rs
@@ -590,25 +590,7 @@ Consider setting 'truncate_ragged_lines={}'."#, polars_error::constants::TRUE)
 
 #[cfg(test)]
 mod test {
-    use super::*;
-
-    #[test]
-    fn test_splitfields() {
-        let input = "\"foo\",\"bar\"";
-        let mut fields = SplitFields::new(input.as_bytes(), b',', Some(b'"'), b'\n');
-
-        assert_eq!(fields.next(), Some(("\"foo\"".as_bytes(), true)));
-        assert_eq!(fields.next(), Some(("\"bar\"".as_bytes(), true)));
-        assert_eq!(fields.next(), None);
-
-        let input2 = "\"foo\n bar\";\"baz\";12345";
-        let mut fields2 = SplitFields::new(input2.as_bytes(), b';', Some(b'"'), b'\n');
-
-        assert_eq!(fields2.next(), Some(("\"foo\n bar\"".as_bytes(), true)));
-        assert_eq!(fields2.next(), Some(("\"baz\"".as_bytes(), true)));
-        assert_eq!(fields2.next(), Some(("12345".as_bytes(), false)));
-        assert_eq!(fields2.next(), None);
-    }
+    use super::SplitLines;
 
     #[test]
     fn test_splitlines() {
