--- crates/polars-io/src/export.rs
@@ -1,4 +1,0 @@
-#[cfg(feature = "parquet")]
-pub use polars_parquet::read::statistics::Statistics as ParquetStatistics;
-#[cfg(feature = "parquet")]
-pub use polars_parquet::read::statistics::*;
