+++ crates/polars-io/src/ipc/mmap.rs
@@ -3,10 +3,53 @@ use arrow::io::ipc::read::{Dictionaries, FileMetadata};
 use arrow::mmap::{mmap_dictionaries_unchecked, mmap_unchecked};
 use arrow::record_batch::RecordBatch;
 use memmap::Mmap;
+use polars_core::frame::ArrowChunk;
+use polars_core::prelude::*;
 
-use super::*;
+use super::ipc_file::IpcReader;
 use crate::mmap::MmapBytesReader;
+use crate::predicates::PhysicalIoExpr;
+use crate::shared::{finish_reader, ArrowReader};
 use crate::utils::{apply_projection, columns_to_projection};
+
+impl<R: MmapBytesReader> IpcReader<R> {
+    pub(super) fn finish_memmapped(
+        &mut self,
+        predicate: Option<Arc<dyn PhysicalIoExpr>>,
+    ) -> PolarsResult<DataFrame> {
+        match self.reader.to_file() {
+            Some(file) => {
+                let mmap = unsafe { memmap::Mmap::map(file).unwrap() };
+                let metadata = read::read_file_metadata(&mut std::io::Cursor::new(mmap.as_ref()))?;
+
+                if let Some(columns) = &self.columns {
+                    let schema = &metadata.schema;
+                    let prj = columns_to_projection(columns, schema)?;
+                    self.projection = Some(prj);
+                }
+
+                let schema = if let Some(projection) = &self.projection {
+                    Arc::new(apply_projection(&metadata.schema, projection))
+                } else {
+                    metadata.schema.clone()
+                };
+
+                let reader = MMapChunkIter::new(mmap, metadata, &self.projection)?;
+
+                finish_reader(
+                    reader,
+                    // don't rechunk, that would trigger a read.
+                    false,
+                    self.n_rows,
+                    predicate,
+                    &schema,
+                    self.row_index.clone(),
+                )
+            },
+            None => polars_bail!(ComputeError: "cannot memory-map, you must provide a file"),
+        }
+    }
+}
 
 struct MMapChunkIter<'a> {
     dictionaries: Dictionaries,
@@ -63,46 +106,6 @@ impl ArrowReader for MMapChunkIter<'_> {
             Ok(Some(chunk))
         } else {
             Ok(None)
-        }
-    }
-}
-
-#[cfg(feature = "ipc")]
-impl<R: MmapBytesReader> IpcReader<R> {
-    pub(super) fn finish_memmapped(
-        &mut self,
-        predicate: Option<Arc<dyn PhysicalIoExpr>>,
-    ) -> PolarsResult<DataFrame> {
-        match self.reader.to_file() {
-            Some(file) => {
-                let mmap = unsafe { memmap::Mmap::map(file).unwrap() };
-                let metadata = read::read_file_metadata(&mut std::io::Cursor::new(mmap.as_ref()))?;
-
-                if let Some(columns) = &self.columns {
-                    let schema = &metadata.schema;
-                    let prj = columns_to_projection(columns, schema)?;
-                    self.projection = Some(prj);
-                }
-
-                let schema = if let Some(projection) = &self.projection {
-                    Arc::new(apply_projection(&metadata.schema, projection))
-                } else {
-                    metadata.schema.clone()
-                };
-
-                let reader = MMapChunkIter::new(mmap, metadata, &self.projection)?;
-
-                finish_reader(
-                    reader,
-                    // don't rechunk, that would trigger a read.
-                    false,
-                    self.n_rows,
-                    predicate,
-                    &schema,
-                    self.row_index.clone(),
-                )
-            },
-            None => polars_bail!(ComputeError: "cannot memory-map, you must provide a file"),
         }
     }
 }
