+++ crates/polars-io/src/parquet/metadata.rs
@@ -3,4 +3,6 @@
 use std::sync::Arc;
 
 pub use polars_parquet::parquet::metadata::FileMetaData;
+pub use polars_parquet::read::statistics::{deserialize, Statistics as ParquetStatistics};
+
 pub type FileMetaDataRef = Arc<FileMetaData>;
