+++ crates/polars-io/src/partition.rs
@@ -1,3 +1,5 @@
+//! Functionality for writing a DataFrame partitioned into multiple files.
+
 use std::fs::File;
 use std::io::BufWriter;
 use std::path::{Path, PathBuf};
@@ -9,22 +11,6 @@ use rayon::prelude::*;
 
 use crate::utils::resolve_homedir;
 use crate::WriterFactory;
-
-/// partition_df must be created by the same way of partition_by
-fn resolve_partition_dir<I, S>(rootdir: &Path, by: I, partition_df: &DataFrame) -> PathBuf
-where
-    I: IntoIterator<Item = S>,
-    S: AsRef<str>,
-{
-    let mut path = PathBuf::new();
-    path.push(resolve_homedir(rootdir));
-
-    for key in by.into_iter() {
-        let value = partition_df[key.as_ref()].get(0).unwrap().to_string();
-        path.push(format!("{}={}", key.as_ref(), value))
-    }
-    path
-}
 
 /// Write a DataFrame with disk partitioning
 ///
@@ -40,7 +26,7 @@ where
 ///         .finish(df)
 /// }
 /// ```
-///
+
 pub struct PartitionedWriter<F> {
     option: F,
     rootdir: PathBuf,
@@ -126,60 +112,18 @@ where
     }
 }
 
-#[cfg(test)]
-mod test {
-    use super::*;
+/// `partition_df` must be created in the same way as `partition_by`.
+fn resolve_partition_dir<I, S>(rootdir: &Path, by: I, partition_df: &DataFrame) -> PathBuf
+where
+    I: IntoIterator<Item = S>,
+    S: AsRef<str>,
+{
+    let mut path = PathBuf::new();
+    path.push(resolve_homedir(rootdir));
 
-    #[test]
-    #[cfg(feature = "ipc")]
-    fn test_ipc_partition() -> PolarsResult<()> {
-        use std::io::BufReader;
-        use std::path::PathBuf;
-
-        use tempfile;
-
-        use crate::ipc::IpcReader;
-        use crate::prelude::IpcWriterOption;
-        use crate::SerReader;
-
-        let tmp_dir = tempfile::tempdir()?;
-
-        let df = df!("a" => [1, 1, 2, 3], "b" => [2, 2, 3, 4], "c" => [2, 3, 4, 5]).unwrap();
-        let by = ["a", "b"];
-        let rootdir = tmp_dir.path().join("ipc-partition");
-
-        let option = IpcWriterOption::new();
-
-        PartitionedWriter::new(option, rootdir.clone(), by).finish(&df)?;
-
-        let expected_dfs = [
-            df!("a" => [1, 1], "b" => [2, 2], "c" => [2, 3])?,
-            df!("a" => [2], "b" => [3], "c" => [4])?,
-            df!("a" => [3], "b" => [4], "c" => [5])?,
-        ];
-
-        let expected: Vec<(PathBuf, DataFrame)> = ["a=1/b=2", "a=2/b=3", "a=3/b=4"]
-            .into_iter()
-            .zip(expected_dfs)
-            .map(|(p, df)| (rootdir.join(p), df))
-            .collect();
-
-        for (expected_dir, expected_df) in expected.iter() {
-            assert!(expected_dir.exists());
-
-            let ipc_paths = std::fs::read_dir(expected_dir)?
-                .map(|e| {
-                    let entry = e?;
-                    Ok(entry.path())
-                })
-                .collect::<PolarsResult<Vec<_>>>()?;
-
-            assert_eq!(ipc_paths.len(), 1);
-            let reader = BufReader::new(polars_utils::open_file(&ipc_paths[0])?);
-            let df = IpcReader::new(reader).finish()?;
-            assert!(expected_df.equals(&df));
-        }
-
-        Ok(())
+    for key in by.into_iter() {
+        let value = partition_df[key.as_ref()].get(0).unwrap().to_string();
+        path.push(format!("{}={}", key.as_ref(), value))
     }
+    path
 }
