+++ crates/polars-io/src/prelude.rs
@@ -1,6 +1,4 @@
-#[cfg(test)]
-use polars_core::prelude::*;
-
+pub use crate::cloud;
 #[cfg(feature = "csv")]
 pub use crate::csv::{read::*, write::*};
 #[cfg(any(feature = "ipc", feature = "ipc_streaming"))]
@@ -11,12 +9,5 @@ pub use crate::json::*;
 pub use crate::ndjson::core::*;
 #[cfg(feature = "parquet")]
 pub use crate::parquet::{metadata::*, read::*, write::*};
+pub use crate::shared::{SerReader, SerWriter};
 pub use crate::utils::*;
-pub use crate::{cloud, SerReader, SerWriter};
-
-#[cfg(test)]
-pub(crate) fn create_df() -> DataFrame {
-    let s0 = Series::new("days", [0, 1, 2, 3, 4].as_ref());
-    let s1 = Series::new("temp", [22.1, 19.9, 7., 2., 3.].as_ref());
-    DataFrame::new(vec![s0, s1]).unwrap()
-}
