+++ crates/polars-lazy/src/physical_plan/executors/scan/parquet.rs
@@ -7,7 +7,8 @@ use polars_core::utils::accumulate_dataframes_vertical;
 use polars_io::cloud::CloudOptions;
 use polars_io::parquet::metadata::FileMetaDataRef;
 use polars_io::parquet::read::materialize_empty_df;
-use polars_io::{is_cloud_url, RowIndex};
+use polars_io::utils::is_cloud_url;
+use polars_io::RowIndex;
 
 use super::*;
