+++ crates/polars-plan/src/logical_plan/functions/count.rs
@@ -13,7 +13,7 @@ use polars_io::parquet::read::ParquetReader;
 #[cfg(all(feature = "parquet", feature = "async"))]
 use polars_io::pl_async::{get_runtime, with_concurrency_budget};
 #[cfg(feature = "parquet")]
-use polars_io::{is_cloud_url, SerReader};
+use polars_io::{utils::is_cloud_url, SerReader};
 
 use super::*;
