+++ crates/polars-plan/src/logical_plan/hive.rs
@@ -95,7 +95,7 @@ impl HivePartitions {
 /// Determine the path separator for identifying Hive partitions.
 #[cfg(target_os = "windows")]
 fn separator(url: &Path) -> char {
-    if polars_io::is_cloud_url(url) {
+    if polars_io::utils::is_cloud_url(url) {
         '/'
     } else {
         '\\'
