+++ crates/polars/tests/it/io/csv.rs
@@ -3,6 +3,7 @@ use std::num::NonZeroUsize;
 
 use polars::io::RowIndex;
 use polars_core::export::chrono;
+use polars_core::utils::concat_df;
 
 use super::*;
 
@@ -133,6 +134,18 @@ fn test_read_csv_file() {
         .unwrap();
 
     assert_eq!(df.shape(), (27, 4));
+}
+
+#[test]
+fn test_read_csv_filter() -> PolarsResult<()> {
+    let df = CsvReader::from_path(FOODS_CSV)?.finish()?;
+
+    let out = df.filter(&df.column("fats_g")?.gt(4)?)?;
+
+    // This fails if all columns are not equal.
+    println!("{out}");
+
+    Ok(())
 }
 
 #[test]
@@ -1246,4 +1259,18 @@ fn test_leading_whitespace_with_quote() -> PolarsResult<()> {
     assert_eq!(col_1.get(0)?, AnyValue::Float64(24.5));
     assert_eq!(col_2.get(0)?, AnyValue::String("  4.1"));
     Ok(())
+}
+
+#[test]
+fn test_read_io_reader() {
+    let path = "../../examples/datasets/foods1.csv";
+    let file = std::fs::File::open(path).unwrap();
+    let mut reader = CsvReader::from_path(path).unwrap().with_chunk_size(5);
+
+    let mut reader = reader.batched_borrowed_read().unwrap();
+    let batches = reader.next_batches(5).unwrap().unwrap();
+    assert_eq!(batches.len(), 5);
+    let df = concat_df(&batches).unwrap();
+    let expected = CsvReader::new(file).finish().unwrap();
+    assert!(df.equals(&expected))
 }
